chore: refactor product seeder
