+++ test/unit/controller/point-of-sale-controller.ts
@@ -24,7 +24,6 @@ import User, { UserType } from '../../../src/entity/user/user';
 import {
   seedContainers,
   seedPointsOfSale,
-  seedProducts,
 } from '../../seed-legacy';
 import TokenMiddleware from '../../../src/middleware/token-middleware';
 import PointOfSale from '../../../src/entity/point-of-sale/point-of-sale';
@@ -61,7 +60,7 @@ import { SwaggerSpecification } from 'swagger-model-validator';
 import { SeededRole } from '../../seed/rbac';
 import PointOfSaleService from '../../../src/service/point-of-sale-service';
 import MemberAuthenticator from '../../../src/entity/authenticator/member-authenticator';
-import { ProductCategorySeeder, RbacSeeder, UserSeeder, VatGroupSeeder } from '../../seed';
+import { ProductSeeder, RbacSeeder, UserSeeder } from '../../seed';
 
 chai.use(deepEqualInAnyOrder);
 
@@ -130,9 +129,7 @@ describe('PointOfSaleController', async () => {
     const feut2 = users.filter((u) => u.type === UserType.MEMBER)[1];
     const bestuur1 = users.filter((u) => u.type === UserType.MEMBER)[2];
 
-    const categories = await new ProductCategorySeeder().seedProductCategories();
-    const vatGroups = await new VatGroupSeeder().seedVatGroups();
-    const { productRevisions } = await seedProducts([adminUser, organUser], categories, vatGroups);
+    const { productRevisions } = await new ProductSeeder().seedProducts([adminUser, organUser]);
     const { containers, containerRevisions } = await seedContainers([adminUser, organUser], productRevisions);
     const { pointsOfSale, pointOfSaleRevisions } = await seedPointsOfSale([adminUser, organUser], containerRevisions);
