+++ test/unit/controller/product-controller.ts
@@ -30,7 +30,6 @@ import path from 'path';
 import sinon from 'sinon';
 import User, { TermsOfServiceStatus, UserType } from '../../../src/entity/user/user';
 import Database from '../../../src/database/database';
-import { seedProducts } from '../../seed-legacy';
 import TokenHandler from '../../../src/authentication/token-handler';
 import Swagger from '../../../src/start/swagger';
 import RoleManager from '../../../src/rbac/role-manager';
@@ -49,7 +48,7 @@ import VatGroup from '../../../src/entity/vat-group';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
 import ProductRevision from '../../../src/entity/product/product-revision';
-import { ProductCategorySeeder, RbacSeeder, VatGroupSeeder } from '../../seed';
+import { ProductSeeder, RbacSeeder, VatGroupSeeder } from '../../seed';
 
 /**
  * Tests if a product response is equal to the request.
@@ -129,9 +128,8 @@ describe('ProductController', async (): Promise<void> => {
     await User.save(organ);
     const users = [organ, adminUser, localUser];
 
-    const categories = await new ProductCategorySeeder().seedProductCategories();
     const vatGroups = await new VatGroupSeeder().seedVatGroups();
-    const { products } = await seedProducts(users, categories, vatGroups);
+    const { products } = await new ProductSeeder().seedProducts(users, undefined, vatGroups);
 
     const validProductReq: UpdateProductRequest = {
       name: 'Valid product',
