+++ test/unit/service/point-of-sale-service.ts
@@ -28,7 +28,6 @@ import Database from '../../../src/database/database';
 import {
   seedContainers,
   seedPointsOfSale,
-  seedProducts,
 } from '../../seed-legacy';
 import Swagger from '../../../src/start/swagger';
 import {
@@ -44,7 +43,7 @@ import PointOfSaleRevision from '../../../src/entity/point-of-sale/point-of-sale
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
 import RbacSeeder, { SeededRole } from '../../seed/rbac';
-import { ProductCategorySeeder, UserSeeder, VatGroupSeeder } from '../../seed';
+import { ProductSeeder, UserSeeder } from '../../seed';
 
 chai.use(deepEqualInAnyOrder);
 
@@ -103,11 +102,9 @@ describe('PointOfSaleService', async (): Promise<void> => {
     await truncateAllTables(connection);
 
     const users = await new UserSeeder().seedUsers();
-    const categories = await new ProductCategorySeeder().seedProductCategories();
-    const vatGroups = await new VatGroupSeeder().seedVatGroups();
     const {
       productRevisions,
-    } = await seedProducts(users, categories, vatGroups);
+    } = await new ProductSeeder().seedProducts(users);
     const {
       containers,
       containerRevisions,
