+++ test/unit/service/product-service.ts
@@ -26,7 +26,7 @@ import Database from '../../../src/database/database';
 import Swagger from '../../../src/start/swagger';
 import ProductService, { ProductFilterParameters } from '../../../src/service/product-service';
 import {
-  seedProducts, seedContainers,
+  seedContainers,
   seedPointsOfSale,
 } from '../../seed-legacy';
 import Product from '../../../src/entity/product/product';
@@ -51,7 +51,7 @@ import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
 import sinon from 'sinon';
 import { ContainerWithProductsResponse } from '../../../src/controller/response/container-response';
-import { ProductCategorySeeder, UserSeeder, VatGroupSeeder } from '../../seed';
+import { ProductCategorySeeder, ProductSeeder, UserSeeder, VatGroupSeeder } from '../../seed';
 
 chai.use(deepEqualInAnyOrder);
 
@@ -147,15 +147,13 @@ describe('ProductService', async (): Promise<void> => {
     const connection = await Database.initialize();
     await truncateAllTables(connection);
 
-    const categories = await new ProductCategorySeeder().seedProductCategories();
-    const vatGroups = await new VatGroupSeeder().seedVatGroups();
     const users = await new UserSeeder().seedUsers();
 
     const {
       products,
       productImages,
       productRevisions,
-    } = await seedProducts(users, categories, vatGroups);
+    } = await new ProductSeeder().seedProducts(users);
     const {
       containers,
       containerRevisions,
