+++ test/unit/service/vat-group-service.ts
@@ -26,14 +26,13 @@ import Database from '../../../src/database/database';
 import {
   seedContainers,
   seedPointsOfSale,
-  seedProducts,
   seedTransactions,
 } from '../../seed-legacy';
 import VatGroupService from '../../../src/service/vat-group-service';
 import { VatDeclarationResponse } from '../../../src/controller/response/vat-group-response';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
-import { ProductCategorySeeder, UserSeeder, VatGroupSeeder } from '../../seed';
+import { ProductSeeder, UserSeeder, VatGroupSeeder } from '../../seed';
 
 describe('VatGroupService', () => {
   let ctx: {
@@ -49,8 +48,7 @@ describe('VatGroupService', () => {
     await truncateAllTables(connection);
     const users = await new UserSeeder().seedUsers();
     const vatGroups = await new VatGroupSeeder().seedVatGroups();
-    const categories = await new ProductCategorySeeder().seedProductCategories();
-    const { productRevisions } = await seedProducts(users, categories, vatGroups, 100);
+    const { productRevisions } = await new ProductSeeder().seedProducts(users, undefined, vatGroups, 100);
     const { containerRevisions } = await seedContainers(users, productRevisions);
     const { pointOfSaleRevisions } = await seedPointsOfSale(users, containerRevisions);
     const { transactions } = await seedTransactions(users, pointOfSaleRevisions, new Date('2020-02-12'), new Date('2022-11-30'), 3);
